Import of PathScale, version 1.0
The PathScale sources have been pre-molded in the following way:
1.  The CVS keywords have been normalized for this repository and do
    not generate any diffs. For filenames this means that they have
    osprey1.0 and not kpro64 as the first level directory even though
    the file lives under kpro64.
2.  The CVS Log keyword in kpro64/arith/edition.c has been removed
    and the delta caused by it due to Fred Chow's initial import has
    been reverted.
3.  The cygnus_20030612 directory has been renamed to cygnus. This
    to reduce diffs with the Pro64 compiler, which had cygnus_20000828
    and that got renamed to cygnus before importing. Future PathScale
    releases have renamed this directory to cygnus as well.

Diff-reduction has been applied in the following way:
1.  Source files that only contained copyright assertions but no other
    (non-whitespace) changes have been reverted. 

This commit does not remove files that are part of Pro64 but are not
part of PathScale 1.0. There are files we do not want to lose, so this
takes some additional effort. Also, no changes have been made yet to
account for the cygnus_20030612->cygnus name change. This will all be
resolved in follow-up commits.



git-svn-id: http://svn.open64.net/svnroot/open64@71 ead328b1-e903-0410-a19d-eacee4b031f6