add a handy flags to diagnose CG bugs -Wb,-SKIP:glos_skip_bb_{e|a|b}=<pu-num>%<bb-num>.
For example, if -Wb,-SKIP:glos_skip_bb_a=3%24 is 
added to the command line, the OPs in blocks whose id > 24 in PU 3
remains unchanged (i.e, no OPs are move into/out-of those blocks, 
and the order of OPs in that blocks are remain unchanged. 
in a nutshell, those blocks are isolated form global scheduling). 




git-svn-id: http://svn.open64.net/svnroot/open64@522 ead328b1-e903-0410-a19d-eacee4b031f6