Implement handling of misaligned long double loads and stores. The algorithm is based on the setf.sig and setf.exp instructions for loads and getf.sig and getf.exp instructions for stores. The bit fiddling achieves the format conversion as described in the SDM, volume 1, figure 5-9: FP register to memory data translation. Also, for loads, the exponent is adjusted for the difference in bias for the register format (bias is 65535) and double-extended memory format (bias is 16383).
For readability purposes, the handling of FP types has been moved
to Expand_Misaligned_Load() and Expand_Misaligned_Store(), which
leaves Expand_Composed_Load() and Expand_Composed_Store() with the
single task of composing the loads and/or stores. Note that misaligned
long double loads and stores are being decomposed into 2 integer loads
and stores. Since the alignment requirements of the integer loads and
stores are less than that of the original double-extended load or store,
it is possible that Expand_Composed_Load() and Expand_Composed_Store()
is being called for *aligned* memory operations. This case is detected
and diverted to Expand_Load() or Expand_Store().

This change also fixes a related bug in the variant of the memory 
operation. Variants for memory operations encode volatility and
alignment. For misaligned memory operations, the alignment was not
encoded as the log2 of the alignment, as the comments in variants.h
made us believe, but rather as the modulo of the required alignment.
An alignment of 7, for example, has been observed. This is bunk.
The fix is to use ffs() to find the largest log2 alignment and to
encode it by adding 1 to make sure that an alignment of 1 does not
yield a zero-valued alignment. This is because of the non-intuitive
use of the alignment bits of the variant. Consequently, there's no
need for an expensive gcd() to determine the actual alignment.

Bug: osprey/58, osprey/59
Reviewed by: Sun Chan



git-svn-id: http://svn.open64.net/svnroot/open64@552 ead328b1-e903-0410-a19d-eacee4b031f6