these fixes are done by jianxin and lixin: 1. bug no. 87 Open64 compiler couldn't identify the branch registers in inline asm statements like "b6" in "asm("" : : : "cc", "b6" )" properly. The objective of this fix is to make the compiler recognizing this kind of registers.
2.bug no. 88
Certain types of asm statements could cause assertion failure during global code 
motion phase in original Open64 compiler. This problem is due to the compiler back-end
marking wrong arcs as candidates for data speculation in dependence graph. 
Actually, They are not. The objective of this fix is to eliminate this kind inconsistency
in the compiler. 

3. bug no.90
he bug occurred in Pre_Optimizer using native compiler. The same error 
may not happen in cross compiler.


git-svn-id: http://svn.open64.net/svnroot/open64@568 ead328b1-e903-0410-a19d-eacee4b031f6