1 This patch is worked by Shuxin and Wenguang.
2 The purpose of the patch is to provide faster long/long and long%long
implementation. GCC's __moddi3 assembly routine is used as a reference.

3. Perforamnce gain
The improvement on parser is 256s vs 270s, roughly 5% improvement. Gap gets
around 2.5% improvement. Improvement on other applications would accumulate
1% improvement on SPEC2K int.

4. Pending issues:

(1) I did not process the condition of "unsigned long". There is a separate
gcc assembly code for it __umoddi3() with just a few instructions different
than the signed version. Also I did not change the code for I4/I2 division
code as Shuxin proposed.
The reason is lack of motivating examples( they can not get speedup in SPEC
2K INT ). While it is worth to do it, I'd rather to check in the current
code to improve the performance of parser because it is more urgent.

(2) div 0
In the current opencc implementation, div 0 won't cause any trouble and we
get wrong results. Will ask Sun how to deal with it.


git-svn-id: http://svn.open64.net/svnroot/open64@697 ead328b1-e903-0410-a19d-eacee4b031f6