Import the GNU4 FE. This FE is based on GCC 4.0.2. Many unnecessary files/directories have been cut off. Only the gcc, libcpp and libiberty and necessary include files and config files are remained. The size of the whole gnu4_fe is 61M.
This check-in will not modify the Makefile and install script. There should be more testing on this folder, especially on X8664.
The Makefile and install script will be updated after finishing testing the new GNU4 FE.

To build the GNU4 FE, please follow the steps below:
$ cd <trunk>/kpro64/gnu4_fe/objdir
On IA64, please run:
$ ./CONFIGURE.ia64
On X8664/IA32, please run:
$ ./CONFIGURE.x8664
Then, build the FE:
$ make
There should be NO any ERRORS during make.
At Last, install the FE manually:
On IA64:
$ cp <trunk>/kpro64/gnu4_fe/objdir/gcc/cc1 $TOOLROOT/usr/lib/gcc-lib/ia64-open64-linux/<OPEN64_VERSION>/
$ cp <trunk>/kpro64/gnu4_fe/objdir/gcc/cc1plus $TOOLROOT/usr/lib/gcc-lib/ia64-open64-linux/<OPEN64_VERSION>/
On X8664/IA32:
$ cp <trunk>/kpro64/gnu4_fe/objdir/gcc/cc1 $OPEN64_INSTALL_PREFIX/lib/<OPEN64_VERSION>/
$ cp <trunk>/kpro64/gnu4_fe/objdir/gcc/cc1plus $OPEN64_INSTALL_PREFIX/lib/<OPEN64_VERSION>/

To use the GNU4 FE, just add '-gnu4' to the command line. For example:
$ opencc -gnu4 helloworld.c

Enjoy the GNU4 FE :-)


git-svn-id: http://svn.open64.net/svnroot/open64@950 ead328b1-e903-0410-a19d-eacee4b031f6