This tiny change is about ltoffx+ldxmov/ld8.mov relocation support. The code sequence of global data will become :     addl r2=@ltoffx(dataname#),gp,     ld8.mov r2=[r2], dataname# // HP cc use ldxmov instead     ld1 r3=[r2]; Linker will change the ld8.mov into 'mov' if the load from GOT turns out to be unnecessary. To the best of my knowledge, this reloc support is dedicatd to ia64. Since it acutally does not entails HW support, I guess similar tech/trick are availale for other arch.
This change is super-trival. I associate the OP loading address from GOT with the corresponding symbol, the code-emit is intercerpted so that the ld8.mov can be emitted properly. Some small benchmarks as well as CPU2k are tested. Crafty obtains 3-4% improvement (I am sure it is genuine). Others remain unchanged (or I am not sure). 


git-svn-id: http://svn.open64.net/svnroot/open64@1107 ead328b1-e903-0410-a19d-eacee4b031f6