Fix OSP_454: the initialize data for complex constant is wrong in 64-bit compiler.
In crayf90/fe90/s_cnstrct.c, it will pack/unpack the two part of the complex
into/from a 64-bit long. But the order is not consistent in packing and unpacking.
In Fortran FE cwh_stab.cxx, the image part of the complex is read from the high
32bit of the real part.

Previous check-in (rev 1978) was applied to the merge08 branch.


git-svn-id: http://svn.open64.net/svnroot/open64@1985 ead328b1-e903-0410-a19d-eacee4b031f6