In bash the implementation of test -e <path> is:
int
unary_test (op, arg)
     char *op, *arg;
{
  intmax_t r;
  struct stat stat_buf;
     
  switch (op[1])
    {
    case 'a':			/* file exists in the file system? */
    case 'e':
      return (test_stat (arg, &stat_buf) == 0);

Note that -a is an (undocumented) alias for -e.

Replace "test -a path" with "test -e path", since treating -a
as alias for -e infix operator is not POSIX compilant (for
example, the alias is not supported in dash's built-in
implementation test).

Code Review: Mei Ye
See bug 15232.



git-svn-id: http://svn.open64.net/svnroot/open64@2417 ead328b1-e903-0410-a19d-eacee4b031f6