Fix Fortran runtime error message output (PCE bug 15182).
Instead of using an external message catalog for Fortran runtime error
messages, include the existing message text from the lib.msg file in the
runtime library.

This change adds a script to extract the messages from lib.msg during the
libu library build and generate a C function that can return the default
message string for a given message identifier.  In _lmessage, we call this
routine to get the raw message text instead of calling catgetmsg.

The verbose message mechanism designed for IRIX systems has never worked
for Open64 on Linux and was therefore removed.



git-svn-id: http://svn.open64.net/svnroot/open64@2427 ead328b1-e903-0410-a19d-eacee4b031f6