Before writing the IR at the end of ipa, under BUILD_OPTIMIZE=DEBUG, we check if the WHIRL nodes form a valid tree. To do this, it maps every node to its parent. If the node is not already visited the map entry should be NULL, else it asserts. In order to map the node to its parent, the node is assigned a free map id (if it does not already have an id) from the free list of the maptable of the current procedure. But it so happens that this id, although in the free list, is being used by another node in the same procedure. When we visit the other node, it appears that this node is already visited as it is now mapped to the first node's parent.
The problem seems to be in partial inlining, where we clone a callee into the
caller and then later delete some of the leftover nodes of the callee from
the caller using WN_DELETE_Tree. When a node is deleted, its map id is added to
the Current_Map_Tab's free list. In this case, a prior call to Set_Tables, 
sets the Current_Map_Tab to the callee's table. So even though this map id is
in use in the callee, it is marked as free in the callee's map table.
Since this node has been cloned into the caller, we need to delete the id 
from the caller's table instead. To do this, we need to explicitly call
IPA_WN_DELETE_Tree, with the caller's map table.





git-svn-id: http://svn.open64.net/svnroot/open64@2443 ead328b1-e903-0410-a19d-eacee4b031f6