Back out previous two change to ldlang.c, that merged changes from Pathscale's ldlang.c which removed the ability of ipa_link to process both WHIRL and normal objects in a normal archive file.
This change still disables the handling of "mixed" archives by
ipa_link by default (real the issue is that the normal linker needs
minor modifications to always ignore archive members that contains
WHIRL when the final link occurs).

The option:

--allow-mixed-objects-in-libraries

has been added.  To pass to ipa_link via the driver specify:

-Wj,--allow-mixed-objects-in-libraries

However if you do specify this flag, you may need to have a modified
linker in your PATH for the final link to succeed.



git-svn-id: http://svn.open64.net/svnroot/open64@2474 ead328b1-e903-0410-a19d-eacee4b031f6