Fix for bug 15260. Under BUILD_OPTIMIZED=DEBUG the compiler asserts on the following code snippet:
static char ca[10] = {0,1,2,3,4,5,6,7,8,9};
char prec1a()
{
      int j = -1;
      return ca[-j];
}

Here is the Whirl dump at the time of the assert, which occurs during alias classification:

I4INTCONST -1 (0xffffffffffffffff)
I4STID 0 <2,1,j> T<4,.predef_I4,4> {freq: 0, ln: 21, col: 0}
   U4LDA 0 <1,50,ca> T<54,anon_ptr.,4>
   I4I4LDID 0 <2,1,j> T<4,.predef_I4,4>
  I4SUB
 I4I1ILOAD 0 T<4,.predef_I4,4> T<55,anon_ptr.,4>
I4CVTL 8
I4STID 1 <1,4,.preg_I4> T<4,.predef_I4,4> # $r1 {freq: 0, ln: 23, col: 0}

The crux of the problem is the fact that the routine ALIAS_CLASSIFICATION::Expr_may_contain_pointer() returns FALSE for the expression rooted at the I4SUB.  This is obviously incorrect because the I4SUB is consumed by an ILOAD.  This problem is exposed by the resolution of a previous issue (r242).

The reason this routine returns false is that in x86_64 and most other targets we set Boolean_type (and Boolean_type2) to be MTYPE_I4, therefore the result type of the subtract, I4, is believed to not be a possible pointer value.   Under -m32 we cannot distinguish a boolean from an I4 that may be a pointer value.

The fix is to remove the MTYPE_is_boolean(res) check from this expression, which allows Expr_may_contain_pointer() to correctly return TRUE.  Under 64-bit mode, MTYPE_I4 < Pointer_size, so the behavior of this routine is unchanged.  This may have some performance implications to -m32 (though likely not large). 


git-svn-id: http://svn.open64.net/svnroot/open64@2479 ead328b1-e903-0410-a19d-eacee4b031f6