Fix for bug 15273: Under OPTIMIZE=DEBUG build the symbol table verifier asserts on the following program
integer :: j(1:0)

j = 1
End

Notice that j(1:0) is a zero-sized array which is allowed by the language.
In such case, there shouldn't be any initializer

FE seems to put an INITO entry for j with an initv table index value as 0.

Symtab verifier asserts that an INITO table entry must have a non-zero INITV table index.

The fix is to recognize zero-sized array symbols and relax the assert for INITO entries for these. 




git-svn-id: http://svn.open64.net/svnroot/open64@2484 ead328b1-e903-0410-a19d-eacee4b031f6