Fix for Bug 15333
This defect was introduced by previous fix for Bug 15273.  Checking for
size(type)==0 is not correct for recognizing zero dimensional arrays.
Fortran zero length character variables can have 0 size and be initialized to "".  The correct fix is to examine the ARB structure of the type associated
with the symbol and recognize a zero dimensional array as one where the 
lower bound and upperbound are constants and lower bound is greater than 
upper bound. 




git-svn-id: http://svn.open64.net/svnroot/open64@2490 ead328b1-e903-0410-a19d-eacee4b031f6