o fix bug 15218: Openf90 - segmentation fault on array reduction in OpenMP
This change fixed IR generation of the reduction for dummy arrays. 
Also, in previous implementation, only constant sized arrays were allowed
in array reduction, which is not conforming to OMP standard. This change 
extended the implemntation to any explicit-shape array. 
For VAR_REDUCTION_ARRAY_OMP type variable, we allocate the local 
version variable in stack.




git-svn-id: http://svn.open64.net/svnroot/open64@2498 ead328b1-e903-0410-a19d-eacee4b031f6