Fixed ABI test test_passing_integers.c and test_basic_returning.c.
In the tests, there are casts from void return function pointer to 
value return function pointer, then the casted function is called and
the return value is assigned to local variable. In be/opt/opt_emit_template.h
the ICALL is changed to CALL since it thought the call is not a real 
indirect function call, but when it converted oo CALL, the return value
is lost. SO the fix is if the type of ICALL is not the same as called
function's type, donot convert the ICALL to CALL.




git-svn-id: http://svn.open64.net/svnroot/open64@2505 ead328b1-e903-0410-a19d-eacee4b031f6