o fix for BUG 15031: astar peak dies with segmentation fault at   run time with "-apo" flag.
This defect actually has nothing to do with apo. The real cause
is that we mark one rodata section (.section .rodata._ZTV9regwayobj)
as writeable, which somehow confuses linker to generate different 
relocate types. The binary core-dumps in loader when dld writes 
relocation address to rodata section.

The root cause for that section not being created as a rodata
section is we did not mark a declaration of "extern const" as
ST_is_const_var when there is no initializer.
Like "extern const int foo;" foo is not marked as const_var.
In IPA when we create global symbol table, we set the symbol 
as const iff all entries are const. In this case, const 
attribute is not set while named section (in the ST_ATTR) 
still says it's a rodata.

The changes are:
(1)In the wgen, if we find a const type to an external variable  
   without an initializer, we'll set that symbol const.
(2)In IPA symbol table merge, we check the named section. If 
   it's read-only, we'll set the global symbol const. This 
   deals with inconsistent const declarations.




git-svn-id: http://svn.open64.net/svnroot/open64@2519 ead328b1-e903-0410-a19d-eacee4b031f6