o fix Bug15368: Invalid machine type V16F8 in Host_TO_Targ in LNO
When vectorizing const, SIMD determines the type of the const by 
looking at the type of the parent of simd_op (istore). When the 
simd_op is a CVT, the logic is wrong. Like in the following case, 
it thinks U4INTCONST as F8CONST, so it applies the vectorized 
type to const_wn directly, which triggers an assert.

  U4INTCONST 4 (0x4)     ### const_wn
 F8I4CVT                 ### simd_op
 F8CONST <1,72,____8.000000000000000>
F8MPY                    ### istore

The fix is to defect CVT or TRUNC and use the const type directly for
these cases. Also remove a dead statement.




git-svn-id: http://svn.open64.net/svnroot/open64@2535 ead328b1-e903-0410-a19d-eacee4b031f6