This is enhancement to eliminate the innermost loop only containing the useless store. It is performed in preopt before LNO. The option -WOPT:useless_store_elimination=false can be used to turn off this transformation. The option -Wb,-tt26:0x1000 can be used to print out the trace for this transformation. This transformation improves 473.astar peak performance by 5-15% depending on number of copies run for peak.
Here is an example for this enhancement:
 for (i=0; i< N;i++) {
  c++;
  if (c == 10000) c = 1;        
  for (j=0; j< M;j++) a[j] = 0; <-- useless store loop
   a[x] = c;
   if (a[y] != c) printf(stdout,"%d\n",y);
 }
free(a);

The only use of a[y] is compared with c (a is freed after loop i). When comparing, a[y] can 
be two values 0 (i.e., !c) or c. In each iteration, c is monotonically increased. So, even 
if the a[j] is not cleared in each iteration, the semantics of (a[y] != c) does not change.  
Since the previous iterations' values are all !c, which have the same  semantics as 0 when 
comparing with c.  However, for the first iteration, some elements of a may have value of c, 
thus a[]  should be cleared. Another case that a needs be cleared is when c is wrapped around. 
In the above example, our transformation is to guard the j loop by the 
condition "if (i==0 || c==1)".





git-svn-id: http://svn.open64.net/svnroot/open64@2536 ead328b1-e903-0410-a19d-eacee4b031f6