This enhancement extends SW prefetching to deal with ARRAY with an inductive base address. This enhancement is currently available only under -LNO:prefetch=3 and can be turned off using -LNO:prefetch_inductive=off.
An "inductive base aadress ARRAY" is defined as the following.
   (1) the all dimensions have constant strides,
   (2) the the base address of the ARRAY is inductive (a linear
expression
       of the innermost loop index var), and
   (3) the stride of the inductive expression is a compile-time
constant.

Following is an example.

   {int loopend; loopend=sites_on_node;
    site *lattice, *s;

     for( i=0, s=lattice ; i<loopend; i++,s++ ){
       register double  ar,br;

       ar = s->tempmat1->e[0][0];
       ...
       br = s->link[dir[j]]->e[2][2];
     }
   }

Both s->tempmat1->e[0][0] and s->link[dir[j]]->e[2][2] meet the above
definition and are treated as ARRAYs with inductive base addresses.

A variant is called indirect inductive base address, which follows
the above definition except of replacing (2) with the following.
   - the base address of the ARRAY contains one indirect load whose
     address
     is inductive and the rest of the base address is loop invariant.

Note that the current volume and locality analyses are too inaccurate
to the inductive base address cases. Hence, while we are reusing the
data structures, such as PF_BASE_ARRAY and PF_UGS, which were
constructed
at PF_LOOPNODE::Process_Refs(), we discard PF_LG constructed earlier and
the analysis from the current framework for the inductive base
address cases. Instead we are using the following framework for the
inductive 
base address cases.

   Santhanam, Gornish, & Hsu, Data Prefetching on the HP PA-8000, ISCA
97.

This is a simple and effective framework, in particular sufficient for
less complicated and overlapped access patterns.

The part that we leverage is where we represent the same partition of
linear inductive address expressions (a1*i + b1 and a2*i + b2, where
a1 = a2) as one PF_LG. We sort refs in the PF_LG based on the
constant offset terms. We then select leaders, where any ref whose
offset within 1 cache line of the previous leader will not be a leader.
We will then insert prefetches only for the leaders.




git-svn-id: http://svn.open64.net/svnroot/open64@2541 ead328b1-e903-0410-a19d-eacee4b031f6