Enhances the capabilities of the existing loop multi-version implementation to provide for handling of:
(a) loops with non-constant upper bounds
(b) references of induction variables beyond the "primary iv"
(c) an arbitrary number of address vectors

In addition, a number of defects in the original implementation were resolved.
The heuristics have been modified, but are still rather targeted in their
candidate selection.  In the future, we plan to extend these heuristics as
opportunties are identified.

Note:
- LNO heuristics are adjusted in the presence of multi-versioned loops.
  Specifically, distribution and peeling of SNLs
  - CG unroll heurstics are adjusted in the presence of multi-version loops.
  - CG branch probabilities for the multi-version precondition check are
    biased towards the loop version with aggressive alias.

Options:
  -WOPT:loop_multiver=[on|off] - defaults to off, but activated under -Ofast
  -WOPT:loop_multiver_aggr=[on|off] - disables heuristics





git-svn-id: http://svn.open64.net/svnroot/open64@2548 ead328b1-e903-0410-a19d-eacee4b031f6