o add a round of memcpy/memset/memmov synthesis in LNO.
current sysnthesis in the lowering phase of mainopt. it does not
work properly as the alias map in alias manager has not been
populated.

the change is under option: -OPT:transform=2 since it flucuates
the performance of a few benchmarks. the synthese in this option
is more aggressive than before.

also fix a few bugs in the original code.




git-svn-id: http://svn.open64.net/svnroot/open64@2558 ead328b1-e903-0410-a19d-eacee4b031f6