A redundancy elimination scheme that takes advantage of the associative property is introduced. This pass is meant to enable the redundancy elimination opportunity identified for cactusADM. The current implementation gets the intended benefit for cactus and maybe does as good a job as optimizing by hand.
Motivating examples:
Open64's PRE or GVN passes are currently not taking 
advantage of the associative property. 
The following example gets optimized by GVN or PRE. 
D = .
G = .
I = 5*G*D
J = 5*G
Subexpression 5*G is eliminated leaving 
only two multiplies in the assembly.
However, when the same example is rewritten 
with a different association, it does not get optimized. 
D = .
G = .
I = 5*G*D
J = 5*D
Neither does the following rewrite get optimized.
D = .
G = .
I = 5*G*D
J = G*D
An earlier canonicalization led to a expression 
tree that looks like this (D*(5*G)) all the time. 
Only 5*G will ever be optimized.

These examples show that we can use reassociation 
in open64. Reassociation can be enabled with a 
subexpression elimination pass that is aware of the 
Math laws for associative arithmetic and then uses these 
laws to guide CSE. Real associative trees like in SPEC 2006 
cactusADM are very large. In such instances we found 
reassociation to introduce performance gain. For cases 
like CactusADM we not only need reassociation for term 
based NARY trees (like the above example) but also for 
trees that also include trees in them, i.e. non-terminal 
based NARY trees, like (A*B+C*D).

We worked on a recursive formulation that works from 
term based and goes up into non-term based trees that 
contain them. We have now implemented a nary-reassociation 
pass. Like earlier noted, cactusADM benefits a lot as a 
result of reassociation enabled redundancy elimination. 
We found that this pass has brought down the number of 
redundant FP adds/subs/mpys for the hottest block in 
cactusADM and enabled a noticeable % improvement in 
performance. 

We first prototyped the algorithm separately from open64. 
Then we incorporated the work into open64. It is a new pass 
which works on basicblocks that contain WHIRL. 

Our intention is to build a standard recursive algorithm 
(using dynamic programming) that selects a subexpression 
elimination candidate by using the number of operations that 
get eliminated as a cost metric. Most of the challenges have 
been in getting an NARY representation of the associative 
input WHIRL trees, which are in the form of binary trees 
for OPR_ADD, OPR_MPY and unary trees for OPR_NEGs and leaf 
WHIRLs for OPR_LDID and OPR_CONSTs. We rewrite WHIRL into 
newly introduced NARY data structures, nary_exp and 
wcexp and work on these NARY forms.

We have added comments in this file, opt_reasso.cxx, that 
should answer a number of questions. Please contact via 
appropriate forum with other questions. 

Placement of pass:
It is placed right after the proactive fusion transformation 
in Pre_optimizer. We invoke it in Pre_optimizer because 
it requires code in form of basic blocks. We decided to 
work with WHIRL with intention to incorporate n-ary 
reassociation in to the PRE/GVN pass in the future. 

Enabling the pass:
The pass is always on, but can be disabled by unsetting 
a flag: WOPT_Enable_Reassociation_CSE. However it is only 
active if Roundoff_Level is greater than ROUNDOFF_ASSOC, 
which is the default case with -Ofast. Finally, it is only 
applied during the MAINOPT_PHASE.
WOPT_Enable_Reassociation_CSE is controlled from command line
using the flag: -WOPT:reasso_cse=1/0 to enable/disable the pass. 

Debug: Traces can be obtained by using the REASSO_DUMP_FLAG 
and REASSO_DUMP_FLAG_DEBUG suboptions to ttWOPT2 
(-Wb,-tt26,0x2000, -Wb,-tt26,0x4000).

Restrictions:
1: Pass works at basic block level only. Maybe that is 
a reasonable choice considering that the algorithmic 
complexity of the solution. 
The algorithmic complexity is around O(n^2). 
2: Then it applies redundancy elimination only to 
associative trees (not really a restriction). 
3: Also associative trees should only use LDIDs and CONSTs 
as leaves (array expressions via OPR_ILOADs are not 
allowed in this version). 
4: Also, it is legal only if the variables in the 
basic block are used in a SSA like way. In the future 
we will either be using available 
expressions or SSA BBs and rebuild the front-end of 
the pass. This is a major restriction now as it is hard to 
encounter SSA like programmer code frequently. 
5: Also we only allow statements like OPR_STIDs, 
OPR_ISTOREs, OPR_COMMENTs, OPR_PRAGMAs, OPR_LABEL 
statements in the basic block. This restriction is also 
set aside for future work. 

However, even with these restrictions we found that 
redundancies in cactusADM's hottest basic block were 
brought down close to what we acheived with hand 
optimization of these redundancies. 

The following are main points from these restrictions: 
pass does not work on array accesses, pass is not 
applied if scalars are written to more than once 
in a basicblock or if a def follows a use in a basic block. 

We believe that these restrictions make this pass 
safe from all aliasing problems except when scalars 
alias with arrays. This can happen in some FORTRAN 
equivalence and C union usage patterns. We will work 
on correcting this known aliasing issue. 

Testing:
We have seen all programs from SPEC compile and run 
successfully and measured gain for cactusADM. We also
noticed that reassociative redundancy elimination may 
help other benchmarks like calculix/gamess.



git-svn-id: http://svn.open64.net/svnroot/open64@2560 ead328b1-e903-0410-a19d-eacee4b031f6