o fix [Bug 15410] New: 32 bit integer code will not link with apo -   missing symbols in libopenmp
use inline assembly instread of buildins since the buildins are
not available in -m32 mode.

as I pointed in the comments, we should have checked if sse
is available since x87 uses a different register. but I did
not do that since it's already broken: in the main(), we 
never check it, and we always ld/st to sse control registger 
even "-mno-sse" is specified. this will be fixed later.




git-svn-id: http://svn.open64.net/svnroot/open64@2566 ead328b1-e903-0410-a19d-eacee4b031f6