Fix Bug 15066:
The problem is in WGEN_add_pragma_to_enclosing_regions, it calls Add_Pragma_To_MP_Regions with 
parent_map set to 0 when it does not want to update the map. This is wrong, since 0 represents 
the dependence map. Indeed WN_MAP_UNDEFINED should be used to represent NULL map.

Another problem happens in Process_Parallel_Region when lowering omp pragma. It tries to process 
exception handling code, while this is C code instead of C++ code. The fix is to add C++ check.




git-svn-id: http://svn.open64.net/svnroot/open64@2596 ead328b1-e903-0410-a19d-eacee4b031f6