o fix for Bug 15175: OMP2001 310.wupwise_m fails during execution in 32 bit mode.
in -m32, we add a fake parameter if the call returns a structure. for  
omp outlined routines, we lower the pu twice with action 
lower_return_val, which results in adding two fake parameters. what we 
see is the callee gets wrong arguments.  
 
the fix is to avoid lower_return_val for omp outlined routines. 




git-svn-id: http://svn.open64.net/svnroot/open64@2616 ead328b1-e903-0410-a19d-eacee4b031f6