Bug 15467 -  visibility problem of hidden inline function
When we are in standalone inliner which is invoked for
single translation unit, all global functions are callable
by other TUs in the same modules (DSO or a.out)
even for inline function  which is not preemptible we
still need to export them since in different .o (in same
module), they can be called from there, and C doesn't require
inline function be defined in every translation unit.

For C++, One Definition Rule requires each inline function 
be defined at every TU using the inline function, so
these inline function can not be called by other .o
in theory. In g++ testsuite, there are several test cases
check if the function be dropped out-of-line.




git-svn-id: http://svn.open64.net/svnroot/open64@2617 ead328b1-e903-0410-a19d-eacee4b031f6