Fix for Bug 15411 : calculix fails at run time when compiled with apo in svn463
When doing unroll and jam for factorization we were incorrectly 
passing the number of loops in the loop nest, instead of the 
depth of the outermost loop.  The output difference was caused because
a variable that required scalar expansion didn't get renamed. 
The problem was exposed by -apo where  more loops are unrolled.  

Summary:    





git-svn-id: http://svn.open64.net/svnroot/open64@2621 ead328b1-e903-0410-a19d-eacee4b031f6