Fix for bug 15158: WOPT segfault for PlumHall tests at +O3
The aux symbols are built in two phases during preopt. First we count the 
symbols and initialize an array of aux symbols. Later we initialize the aux
symbols by traversing the whirl tree in Convert_ST_to_AUX. But during the 
tree walk we try to apply the simplifier which could optimize away some of the
symbols and hence their aux symbols are never initialized. So later phases need
to ignore any uninitialized aux symbols.

This checkin also fixes a bug in the simplifier where when simplifying 
ILOADS to an INTCONST, make sure the const_val field in WN is set correctly





git-svn-id: http://svn.open64.net/svnroot/open64@2628 ead328b1-e903-0410-a19d-eacee4b031f6