Bug 15158 also exposed a bug in scalar expand where we were not checking if the def/use list is null.
Interestingly, this exposed a 15% degradation in calculix, as this check
now prevented some loops from being vectorized. It appears that there 
was a fundamental bug in the LNO's dependence analysis.
Prior to Analyse_Dependence, we clone the innerloop and perform the 
dependence analysis on the clone. But when cloning the loop, 
we also need to update the Du-Ud chains for the cloned WNs.

In addition, we need to use the newly constructed reduction manager for the
cloned loop when calling 
Generate_Scalar_Dependence_For_Statement_Dependence_Graph.




git-svn-id: http://svn.open64.net/svnroot/open64@2630 ead328b1-e903-0410-a19d-eacee4b031f6