Bug fix 15458:
When generating array reduction in lower phase, the ST type can be KIND_POINTER 
(allocable array and automatic array) and KIND_ARRAY (static array). The problem is 
handling for the KIND_ARRAY case is missing.

Another problem is in fortran FE. When processing reduction clause, if the reduction 
is an automatic array, its base should also be marked as reduction. 




git-svn-id: http://svn.open64.net/svnroot/open64@2637 ead328b1-e903-0410-a19d-eacee4b031f6