This change fixes the 403.gcc apo -m32 runtime output difference error (bug 15383).
The root cause of this defect is invalid placement of spill code in GRA.

When we move a restore out of a LRANGE, we check two conditions:
1. There is a predecessor block in the LRANGE, and
2. All the immediate successors of each predecessor block outside the
LRANGE are in <lunit>'s LRANGE (routine Has_Sucessor_Not_In_LRANGE()).

If we only perform restore placements, everything works fine: The
restore in <lunit> is restore_above. Since all the successors of the
predecessor are in LRANGE. The newly added live-range from the
predecessor to each of its successors does not matter and will not
interfere with any other TN. We can get more freedom for placement than
the original condition 2 described in comments (all the predecessor
blocks out of the LRANGE have one successor).
 
But if we do this together with moving spill out of LRANGE, it becomes
problematic. The condition for spill movement is
1. There is a successor block in the LRANGE, and
2. All the immediate predecessors of each successor block outside the
LRANGE are in <lunit>'s LRANGE.
3. ... <irrelevant>
With these conditions, we place the spill code in the beginning of BB
(spill_above).  The register may be trashed by live-ranged added by
restore placement. 

An example, if we have
BB689: spill GTN2375 (rbx); outedge to  BB690 and BB691
BB690: restore GTN2304 (rbx); outedge to BB691 and BB707
BB691: spill GTN2375 (rbx); outedge to BB692 and BB707
BB707: restore GTN2304 (rbx)
When moving GTN2304 out of BB707, we push the restore back to BB690. In
the next step, when we move GTN2375 out of BB689, we puth it to BB690. 
In BB690 rbx is trashed by GTN2304. Thus we have a bad rbx in BB691, if
taking BB690->BB691 path.

There can be two fixes. A simpler one is to restore the original
condition 2 (all the predecessor blocks out of the LRANGE have one
successor). Another way to fix is to test if the restores added by
Move_Restore_Out_Of_LRANGE() trash the same register when we moving a
spill down. This gives more freedom for placement. The change in this
check-in uses the second method.





git-svn-id: http://svn.open64.net/svnroot/open64@2642 ead328b1-e903-0410-a19d-eacee4b031f6