o this fixes the debug assert at line 923 of ../../be/opt/opt_verify.cxx when compiling PlumHall (bug15265).
When propagating vsyms after DSE, we update the chi opnd with the latest
non-dead sym, even for a dead chi. This is problematic when a dead chi
is resurrect (for iload-lda folding) because the resurrect chi is not
propagated to all its original uses. Combining these two things, we end
up with a live-range overlap for the chi opnd. The fix is to not to 
update the chi opnd for a dead chi in DSE.





git-svn-id: http://svn.open64.net/svnroot/open64@2648 ead328b1-e903-0410-a19d-eacee4b031f6