bug 15476   Open64 bool return value ABI is not compatible with gcc
The problem of this defect is we don't do the sign or zero extension properly for return value type size less than 4. Actually we don't have the type information in the return value. The first step to fix it is to set the whirl node with proper return value type. Then we need to expand the whirl node to use sign or zero extension instruction based on the return type.

ALSO ADDED a debug trace mechanism for TN node. After the cg.so is loaded, set break point at "gdb_stop_here()", and use command "p set_trace_tn(num)" to trace the creation of TNnum, when the traced TN is created, the debugger stops at gdb_stop_here.




git-svn-id: http://svn.open64.net/svnroot/open64@2655 ead328b1-e903-0410-a19d-eacee4b031f6