Bug Fix 15418: (DEBUG) games assertion in lno dependence graph sanity check.
The problem is that dependence is not updated after fully unroll short loop following Peel_2D_Triangle_Loops. 
However, it is difficult to update dependence because these phases are inside SNL_Transform. My fix is 
to mark the loops that are triangle_peeled and let the later Unroll_before_Factorize handle the unrolling.

Also fix a bug in unroll_before_factorize: add a check to see whether the current loop is 
already unrolled (and thus deleted) in the previous round of unrolling.





git-svn-id: http://svn.open64.net/svnroot/open64@2659 ead328b1-e903-0410-a19d-eacee4b031f6