Bug 15514 gnuchess_5.07 failed at -m32 -g and -m32 -O0 in DBG                     builds (svn 571) Bug 15515 Lang.CRI.-m32.O3 output difference error
The fix in svn r571 changed the .return_preg type to the same as corresponding function return type, so if the function return bool, the .return_preg type is set to U1 (%al). Which caused problem in later phase when it tries to create pair of TN from %al. To avoid the problem, I changed the .return_preg type back to U4 or U8, but still generate a CVTL if the load source and target size mismatched, based on the result type and descriptor type of the operation.




git-svn-id: http://svn.open64.net/svnroot/open64@2663 ead328b1-e903-0410-a19d-eacee4b031f6