The following changes resolve a number of issues uncovered while building the mysql application under -O2.  All of the issues uncovered are related to unsigned arithmetic and how the backend, wopt and cg, treat (or ignore) conversions from U4 to U8.
1) The strength reduction transformation does not correctly honor the 
modulo behavior of unsigned integer arithmetic when promoting U4 types to 
U8/I8. As a solution, we prevent further strength reduction of SR 
temporaries involving U4 to U8/I8 conversions be they explicit or implicit.  
This guard for explicit conversions already exists in STR_RED::Is_cvt_linear 
for all targets except X8664.  The same guard is promoted to 
STR_RED::Is_implicit_cvt_linear to catch the case of implicit widening of U4 
types.

2) When entering SSA performing constant folding across explicit or implicit
unsigned conversions is incorrect in general.
o In CODEMAP::Add_expr we prevent folding constants across unsigned 
  conversions in order to preserve modulo behavior of unsigned arithmetic 
o In CODEMAP::Canon_mpy we prevent the folding of constants across unsigned 
  conversions for both explict and implicit conversions

3) Lowering. In the final lowering of SELECT operations, we were inserting 
conversion operations when the result type of the SELECT operands does not 
match the result type of the select. This can cause the introduction of *sign* 
extensions when a zero extension was implied.  We preserve the same behavior 
for Fortran PUs and for X8664 targets we only allow the introduction of zero 
extensions for unsigned conversions under the assumption that if the conversion 
is implicit, it must be a zero extend for this target.

4) Code generation
o When materializing an explicit U4U8CVT operation use a TOP_movzlq instead of
  a TOP_mov32 to prevent the copy from being eliminated as unnecessary
o When optimizing lea operations we do not want to fold address computation
  across a type conversion.





git-svn-id: http://svn.open64.net/svnroot/open64@2665 ead328b1-e903-0410-a19d-eacee4b031f6